+++ test/dummy/app/models/atom/page_reference_with_text.rb
@@ -4,7 +4,7 @@ module Atom
   class PageReferenceWithText < Folio::Atom::Base
     STRUCTURE = {
       content: :string,
-      model: [Folio::Page],
+      model: %w[Folio::Page],
     }
   end
 end